// multiplatform-RSD/api/generated/schemas/activityInputType.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Travel Planner API
 * API documentation for the Travel Planner application
 * OpenAPI spec version: 1.0.0
 */

export type ActivityInputType = typeof ActivityInputType[keyof typeof ActivityInputType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ActivityInputType = {
  TRANSPORT: 'TRANSPORT',
  FLIGHT: 'FLIGHT',
  ACCOMMODATION: 'ACCOMMODATION',
  FOOD: 'FOOD',
  REMINDER: 'REMINDER',
  OTHER: 'OTHER',
} as const;